fix(chat): separate CLI report parsing from React components
